requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
autoawq==0.2.7.post3
autocommand==2.2.2
backports.tarfile==1.2.0
bitsandbytes==0.45.0
certifi==2024.12.14
charset-normalizer==3.4.1
coloredlogs==15.0.1
datasets==3.2.0
dill==0.3.8
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
gekko==1.2.1
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.5
MarkupSafe==3.0.2
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
optimum==1.23.3
packaging==24.2
pandas==2.2.3
peft==0.14.0
platformdirs==4.2.2
propcache==0.2.1
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==18.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rouge==1.0.1
safetensors==0.4.5
sentencepiece==0.2.0
setuptools==75.6.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
tomli==2.0.1
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
triton==3.1.0
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wheel==0.43.0
xformers==0.0.29
xxhash==3.5.0
yarl==1.18.3
zipp==3.19.2
zstandard==0.23.0
